[IMP] *_payment_*,website_*: move transaction state messages to payment method
The job of transaction state messages is to show a message on top of paid documents on the portal. For example, if a confirmed transaction is linked to a sales order, the following message is shown at the top of the page: "Your payment has been successfully processed, thank you!".

Currently, these messages are configurable per provider, but it's unlikely that one will want Stripe to display something different than Adyen for a given transaction state. However, we definitively want to display a different message depending on the transaction's payment method. For example, wire transfer payments use the pending_msg field to display the payment instructions; payment methods like Karna are asynchronous by design, and we'll want to display the appropriate message regardless of the provider used to process the payment; etc. This will also allow modules like payment_custom to work with a single provider and multiple payment methods rather than multiple providers and one payment method each.

task-2941756
