+++ addons/account_payment/views/account_portal_templates.xml
@@ -119,8 +119,8 @@
         <xpath expr="//a[hasclass('close')]" position="after">
             <t t-if="success == 'pay_invoice'">
                 <t t-set="tx_sudo" t-value="invoice.get_portal_last_transaction()"/>
-                <span t-if="tx_sudo.provider_id.sudo().done_msg" t-out="tx_sudo.provider_id.sudo().done_msg"/>
-                <div t-if="tx_sudo.provider_id.sudo().pending_msg and tx_sudo.provider_code == 'custom' and invoice.ref">
+                <span t-if="tx_sudo.payment_method_id.sudo().done_msg" t-out="tx_sudo.payment_method_id.sudo().done_msg"/>
+                <div t-if="tx_sudo.payment_method_id.sudo().pending_msg and tx_sudo.provider_code == 'custom' and invoice.ref">
                     <b>Communication: </b><span t-out='invoice.ref'/>
                 </div>
             </t>
