+++ addons/payment/data/payment_method_data.xml
@@ -2759,6 +2759,9 @@
                          ref('base.EUR'),
                      ])]"
         />
+        <field name="pending_msg" type="html">
+            <p>Please make a unique payment transfer to confirm the SEPA mandate.</p>
+        </field>
     </record>
 
     <record id="payment_method_shopback" model="payment.method">
