+++ addons/payment/views/payment_form_templates.xml
@@ -316,10 +316,10 @@
                 </div>
             </div>
             <!-- === Help message === -->
-            <div t-if="not is_html_empty(provider_sudo.pre_msg)"
+            <div t-if="not is_html_empty(pm_sudo.pre_msg)"
                  class="w-100 mb-0 ms-4 small text-600"
             >
-                <t t-out="provider_sudo.pre_msg"/>
+                <t t-out="pm_sudo.pre_msg"/>
             </div>
         </div>
         <!-- === Inline form === -->
