+++ addons/payment/views/portal_templates.xml
@@ -281,20 +281,20 @@
         </t>
         <t t-elif="tx.state == 'pending'">
             <t t-set="alert_style">warning</t>
-            <t t-set="status_message" t-value="tx.provider_id.sudo().pending_msg"/>
+            <t t-set="status_message" t-value="tx.payment_method_id.sudo().pending_msg"/>
         </t>
         <t t-elif="tx.state == 'authorized'">
             <t t-set="alert_style">success</t>
-            <t t-set="status_message" t-value="tx.provider_id.sudo().auth_msg"/>
+            <t t-set="status_message" t-value="tx.payment_method_id.sudo().auth_msg"/>
         </t>
         <t t-elif="tx.state == 'done'">
             <t t-set="alert_style">success</t>
             <t t-set="status_heading" t-value="'Thank you!'"/>
-            <t t-set="status_message" t-value="tx.provider_id.sudo().done_msg"/>
+            <t t-set="status_message" t-value="tx.payment_method_id.sudo().done_msg"/>
         </t>
         <t t-elif="tx.state == 'cancel'">
             <t t-set="alert_style">danger</t>
-            <t t-set="status_message" t-value="tx.provider_id.sudo().cancel_msg"/>
+            <t t-set="status_message" t-value="tx.payment_method_id.sudo().cancel_msg"/>
         </t>
         <t t-elif="tx.state == 'error'">
             <t t-set="alert_style">danger</t>
