+++ addons/payment_custom/models/__init__.py
@@ -1,4 +1,5 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
+from . import payment_method
 from . import payment_provider
 from . import payment_transaction
