+++ addons/payment_custom/views/payment_provider_views.xml
@@ -15,17 +15,6 @@
             <group name="payment_followup" position="attributes">
                 <attribute name="invisible">code == 'custom'</attribute>
             </group>
-            <field name="pending_msg" position="after">
-                <div class="o_row" colspan="2"
-                     invisible="custom_mode != 'wire_transfer'">
-                    <button string=" Reload Pending Message"
-                            type="object"
-                            name="action_recompute_pending_msg"
-                            class="oe_link ms-0 ps-0"
-                            icon="fa-refresh"
-                            />
-              </div>
-            </field>
         </field>
     </record>
