+++ addons/website_event_sale/tests/test_frontend_buy_tickets.py
@@ -96,7 +96,6 @@ class TestUi(HttpCaseWithUserDemo, TestWebsiteEventSaleCommon):
             'state': 'enabled',
             'is_published': True,
         })
-        transfer_provider._transfer_ensure_pending_msg_is_set()
 
         self.start_tour("/", 'event_buy_tickets', login="admin")
 
@@ -109,7 +108,6 @@ class TestUi(HttpCaseWithUserDemo, TestWebsiteEventSaleCommon):
             'state': 'enabled',
             'is_published': True,
         })
-        transfer_provider._transfer_ensure_pending_msg_is_set()
 
         #  Ensure the use of USD (company currency)
         self.env['product.pricelist'].create({'name': "Public Pricelist"})
@@ -125,7 +123,6 @@ class TestUi(HttpCaseWithUserDemo, TestWebsiteEventSaleCommon):
             'state': 'enabled',
             'is_published': True,
         })
-        transfer_provider._transfer_ensure_pending_msg_is_set()
 
         self.start_tour("/", 'event_buy_last_ticket')
