+++ addons/website_sale_picking/data/payment_method_data.xml
@@ -9,6 +9,9 @@
         <field name="support_tokenization">False</field>
         <field name="support_express_checkout">False</field>
         <field name="support_refund"></field>
+        <field name="pending_msg" type="html">
+            <p><i>Your order has been saved.</i> Please come to the store to pay for your products.</p>
+        </field>
     </record>
 
 </odoo>
