+++ addons/website_sale_picking/data/payment_provider_data.xml
@@ -14,11 +14,6 @@
         />
         <field name="image_128" type="base64" file="website_sale_picking/static/description/icon.png"/>
         <field name="redirect_form_view_id" ref="payment_custom.redirect_form"/>
-        <field name="pending_msg" type="html">
-            <p>
-                <i>Your order has been saved.</i> Please come to the store to pay for your products
-            </p>
-        </field>
     </record>
 
 </odoo>
